[gen] Init def of series
- series have name, tags, default not points
- int and double are handled separatedly since golang does not have a good generic model and it's easier when serialze
- TODO: have some doubt about multiple values, which is supported in InfluxDB and druid but not in KairosDB @czheo
